dev-requirements.txt
piptools==7.3.0
black==23.11.0
django-debug-toolbar==4.2.0